# docs/swagger.yaml
openapi: 3.0.2
info:
  title: barBank
  version: 1.0.0
  contact: {}
  license:
    name: Mozilla 2.0
    url: 'https://www.mozilla.org/en-US/MPL/2.0/'
paths:
  /users:
    post:
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: string
            examples:
              POST /users:
                value:
                  name: Steam
                  username: steamuser
                  password: novichok
        required: false
      tags:
        - users
      responses:
        '201':
          description: "\r\nUser successfully registered."
        '400':
          content:
            application/json:
              schema:
                type: string
              examples:
                Invalid password:
                  value:
                    error: Invalid password
                Parameter is required:
                  value:
                    error: 'User validation failed: .*: Path `.*` is required'
          description: Bad request
        '409':
          content:
            application/json:
              schema:
                type: string
              examples:
                Username already exists:
                  value:
                    error: Username already exists
          description: Username already exists
      deprecated: false
      summary: New user register
  /sessions:
    summary: Create a session
    post:
      requestBody:
        description: A new `session` creation.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/session'
            examples:
              POST /sessions request body:
                value:
                  username: steamuser
                  password: novichok
        required: true
      tags:
        - sessions
      responses:
        '201':
          content:
            application/json:
              examples:
                201 response:
                  value:
                    token: 615bff2c521653b4a1969ab2
          description: Successful response.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid credentials:
                  value:
                    error: Invalid credentials
          description: Invalid credentials
      operationId: createsession
      summary: Create a session
    delete:
      tags:
        - sessions
      responses:
        '204':
          description: Successful logout
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header:
                  value:
                    error: Invalid Authorization header
                Invalid token:
                  value:
                    error: Invalid token
                Invalid authorization header format:
                  value:
                    error: Invalid authorization header format
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
          description: Invalid Authorization header
      security:
        -
          Baerer: []
  /users/current:
    get:
      tags:
        - users
      responses:
        '200':
          description: Succesful request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
                Invalid token:
                  value:
                    error: Invalid token
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header format
          description: Unauthorized
      security:
        -
          Baerer: []
  /transactions:
    summary: Path used to manage the list of transactions.
    description: >-
      The REST endpoint/path used to list and create zero or more `transaction` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    get:
      tags:
        - transaction
      responses:
        '200':
          content:
            application/json: {}
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
                Invalid token:
                  value:
                    error: Invalid token
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header format
          description: Unauthorized
      security:
        -
          Baerer: []
    post:
      requestBody:
        description: A new `transaction` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transaction'
            examples:
              POST /transaction:
                value:
                  accountFrom: 1b784bb50b65304f534480390cc90aafcbc
                  accountTo: 1b788818551170b598016f5c7abf7b16891
                  amount: 200
                  currency: EUR
                  explanation: a
        required: true
      tags:
        - transaction
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid amount:
                  value:
                    error: Invalid amount
                Missing parameter:
                  value:
                    error: 'Transaction validation failed: .* : Path `.*` is required.'
                Parameter shorter than the minimum allowed length:
                  value:
                    error: Path `.*` (`.*`) is shorter than the minimum allowed length
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
                Invalid token:
                  value:
                    error: Invalid token
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header format
          description: Unauthorized
        '402':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Insufficient funds:
                  value:
                    error: Insufficient funds
          description: Insufficient funds
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                accountFrom not found:
                  value:
                    error: accountFrom not found
          description: Not found
      security:
        -
          Baerer: []
      operationId: createtransaction
      summary: Create a transaction
      description: Creates a new instance of a `transaction`.
components:
  schemas:
    session:
      description: ''
      type: object
    user:
      description: ''
      required: []
      type: object
      properties:
        username:
          description: ''
          type: string
        password:
          description: ''
          type: string
        name:
          description: ''
          type: string
      example:
        name: Steam
        username: steamuser
        password: novichok
    error:
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: error message
    transaction:
      title: Root Type for transaction
      description: ''
      type: object
      properties:
        accountFrom:
          type: string
        accountTo:
          type: string
        amount:
          format: int32
          type: integer
        explanation:
          type: string
      example:
        accountFrom: RU000000001
        accountTo: BY000000001
        amount: 1500000000
        explanation: Explanation
  securitySchemes:
    Baerer:
      scheme: bearer
      type: http